ai-hr/services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
requests==2.32.3
PyYAML==6.0.1
